requirements.txt
pandas
numpy
ta
python-binance
backtesting
requests
